requirements.txt
llama-cpp-python>=0.3.0
huggingface-hub>=0.25.0
fastapi>=0.115.0
uvicorn>=0.32.0
requests>=2.31.0
openai>=1.0.0
